Добавлен docker-compose, конфигурация для app, nginx, RabbitMQ и Postgres, структура проекта и .gitignore (задание урока 2)
