+++ Dockerfile
@@ -1,0 +1,16 @@
+FROM python:3.12-slim
+
+RUN apt-get update && apt-get install -y curl && apt-get clean
+
+WORKDIR /app
+
+COPY app/requirements.txt .
+
+RUN pip install --upgrade pip && pip install -r requirements.txt
+
+COPY app/ .
+COPY app/main.py .
+
+EXPOSE 8000
+
+CMD ["uvicorn", "main:app", "--host", "0.0.0.0", "--port", "8000", "--reload"]
