+++ app/.env
@@ -1,0 +1,11 @@
+APP_ENV=dev
+APP_PORT=8000
+
+POSTGRES_DB=spanola
+POSTGRES_USER=postgres
+POSTGRES_PASSWORD=postgres
+POSTGRES_HOST=postgres
+POSTGRES_PORT=5432 
+
+RABBITMQ_DEFAULT_USER=guest
+RABBITMQ_DEFAULT_PASS=guest
