+++ web-proxy/nginx.conf
@@ -1,0 +1,20 @@
+events {}
+
+http {
+  upstream app {
+    server app:8000;
+  }
+
+  server {
+    listen 80;
+    server_name _;
+
+    location / {
+      proxy_pass         http://app;
+      proxy_set_header   Host $host;
+      proxy_set_header   X-Real-IP $remote_addr;
+      proxy_set_header   X-Forwarded-For $proxy_add_x_forwarded_for;
+    }
+  }
+
+}
